requirements.txt
fastapi
uvicorn[standard]
playwright
requests
pydantic
python-dotenv
pandas
pdfplumber
python-multipart